refactor: modularize MCP server for better maintainability
- Extract API client functionality into separate NarrativeApiClient class
- Create dedicated ToolHandlers and ResourceHandlers modules
- Add configuration module for environment validation
- Reduce main index.ts from 357 to 85 lines
- Improve code organization and single responsibility principle
- Maintain all existing functionality and test compatibility
